--- lab5/Matrix/main.cpp
@@ -1,17 +1,0 @@
-//
-// Created by ASUSTek on 2017-04-02.
-//
-
-#include "Matrix.h"
-
-using std::cout;
-using std::endl;
-
-namespace algebra {
-
-    int main() {
-        cout << "XX";
-        cout << "cokolwiek";
-    return 0;
-    }
-}
