+++ lab5/matrix/CMakeLists.txt
@@ -4,7 +4,7 @@ project(matrix)
 set(HEDEAR_FILES Matrix.h)
 set(SOURCE_FILES Matrix.cpp)
 
-add_library(libmatrix STATIC ${SOURCE_FILES} main.cpp)
+add_library(libmatrix STATIC ${SOURCE_FILES})
 target_include_directories(libmatrix PUBLIC .)
 
 add_executable(matrix ${SOURCE_FILES} main.cpp)
