+++ lab5/matrix/Matrix.h
@@ -1,5 +1,5 @@
 //
-// Created by ASUSTek on 2017-04-02.
+// Created by ASUSTek on 2017-04-05.
 //
 
 #ifndef JIMP_EXERCISES_MATRIX_H
@@ -7,33 +7,31 @@
 
 #include <iostream>
 #include <complex>
-#include <string>
-
 namespace algebra {
+
     class Matrix {
     public:
-        Matrix(); //konstruktor bezparametryczny
+        Matrix(); // konstruktor bezparametryczny
 
-        Matrix(int n_rows, int n_cols); //konstruktor parametryczny rozmiar
-
-        Matrix(std::string MatLab); //konstruktor parametryczny matlab
+        ~Matrix();
 
         Matrix(const Matrix &matrix);
 
-        ~Matrix();
+        Matrix(std::string MatLab); //parametryczny matlab
+
+        Matrix(int n_rows, int n_cols); //parametryczny rows x cols
 
 
         std::complex<double> GetElem(int n_rows, int n_cols);
 
         void SetElem(int n_rows, int n_cols, std::complex<double> element);
 
-
     private:
-        int n_rows_;
-        int n_cols_;
         std::complex<double> **matrix_;
-
+        int n_cols_, n_rows_;
     };
+
 }
 
 #endif //JIMP_EXERCISES_MATRIX_H
+
