+++ lab5/matrix/main.cpp
@@ -1,0 +1,11 @@
+//
+// Created by ASUSTek on 2017-04-05.
+//
+
+#include "Matrix.h"
+
+int main()
+{
+    std::cout << "HEHE";
+    return 0;
+}
